src/requirements.txt
alpaca-py
boto3